requirements.txt
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
discord.py==2.3.2
python-dotenv==1.0.0
colorama==0.4.6
click==8.1.7
pyyaml==6.0.1
fake-useragent==1.4.0
schedule==1.2.0
rich==13.7.0
httpx==0.25.2
